Added required Changes And also added loggers.

+++ PensionManagementBankingService/Controller/BankingController.cs
@@ -12,11 +12,14 @@ namespace PensionManagementBankingService.Controller
     [ApiController]
     public class BankingController : ControllerBase
     {
+        
         private readonly IBankingRepository _bankingRepository;
+        private readonly ILogger<BankingController> _logger;
 
-        public BankingController(IBankingRepository bankingRepository)
+        public BankingController(IBankingRepository bankingRepository, ILogger<BankingController> logger)
         {
             _bankingRepository = bankingRepository;
+            _logger = logger;
         }
         [HttpGet]
         public async Task<ActionResult<IEnumerable<BankResponse>>> GetAllBankingDetails()
@@ -24,25 +27,22 @@ namespace PensionManagementBankingService.Controller
             try
             {
                 var bankingDetails = await _bankingRepository.GetAllBankingDetails();
+                if(bankingDetails.Count() == 0)
+                {
+                    throw new BankingExceptions( "BankingDetails Not Found");
+
+                }
+                _logger.LogInformation("Retrieved all banking details successfully");
                 return Ok(bankingDetails);
             }
             catch (BankingExceptions ex)
             {
-                switch (ex.Type)
-                {
-                    case BankingExceptions.ErrorType.DuplicateRecord:
-                        return Conflict(ex.Message);
-                    case BankingExceptions.ErrorType.EmptyResult:
-                        return NotFound(ex.Message);
-                    case BankingExceptions.ErrorType.NotFound:
-                        return NotFound(ex.Message);
-                    default:
-                        throw;
-                }
+                _logger.LogError(ex, "Error occurred while retrieving all banking details");
+                return StatusCode(404,ex.Message);
             }
             catch (Exception ex)
             {
-                
+                _logger.LogError(ex, "An expected error occured while retrieving all bank details");
                 return StatusCode(500, "An unexpected error occured,please try later");
             }
         }
@@ -54,26 +54,20 @@ namespace PensionManagementBankingService.Controller
                 var bankingDetails = await _bankingRepository.GetBankingDetailsById(bankId);
                 if(bankingDetails == null)
                 {
-                    return NotFound();
+                    _logger.LogWarning($"No banking details found for ID: {bankId}");
+                   throw new BankingExceptions("Banking details not found for given Id");
                 }
+                _logger.LogInformation($"Retrieved BankingDetails for ID: {bankId}");
                 return Ok(bankingDetails);
             }
             catch (BankingExceptions ex)
             {
-                switch (ex.Type)
-                {
-                    case BankingExceptions.ErrorType.DuplicateRecord:
-                        return Conflict(ex.Message);
-                    case BankingExceptions.ErrorType.EmptyResult:
-                        return NotFound(ex.Message);
-                    case BankingExceptions.ErrorType.NotFound:
-                        return NotFound(ex.Message);
-                    default:
-                        throw;
-                }
+                _logger.LogError(ex, $"Error occurred while retrieving banking details for given Id{bankId}");
+                return StatusCode(404, ex.Message);
             }
             catch (Exception ex)
             {
+                _logger.LogError(ex, "An expected error occured while retrieving bank details by Id");
                 return StatusCode(500, "An unexpected error occured, please try later");
             }
         }
@@ -82,6 +76,11 @@ namespace PensionManagementBankingService.Controller
         {
             try
             {
+                var existingRecord = await _bankingRepository.GetBankDetailsByPensionerId(bankingDetails.PensionerId);
+                if(existingRecord!=null)
+                {
+                    throw new BankingExceptions("Duplicate record already exist");
+                }
                 BankingDetails bankDTO = new BankingDetails
                 {
                     BankName = bankingDetails.BankName,
@@ -92,6 +91,7 @@ namespace PensionManagementBankingService.Controller
                     PensionerId = bankingDetails.PensionerId,
 
                 };
+                
                 var addedBankingDetails = await _bankingRepository.AddBankingDetails(bankDTO);
                 BankResponse bank = new BankResponse
                 {
@@ -103,24 +103,17 @@ namespace PensionManagementBankingService.Controller
                     PanNumber = addedBankingDetails.PanNumber,
                     PensionerId = addedBankingDetails.PensionerId,
                 };
+                _logger.LogInformation("Banking details added successfully");
                 return Ok(bank);
             }
             catch (BankingExceptions ex)
             {
-                switch (ex.Type)
-                {
-                    case BankingExceptions.ErrorType.DuplicateRecord:
-                        return Conflict(ex.Message);
-                    case BankingExceptions.ErrorType.EmptyResult:
-                        return NotFound(ex.Message);
-                    case BankingExceptions.ErrorType.NotFound:
-                        return NotFound(ex.Message);
-                    default:
-                        throw;
-                }
+                _logger.LogError(ex, "Banking exception occurred while adding banking details");
+                return StatusCode(500,ex.Message);
             }
             catch (Exception ex)
             {
+                _logger.LogError(ex, "An expected error occured while adding bank details");
                 return StatusCode(500, "An unexpected error occured, please try later");
             }
         }
@@ -142,56 +135,49 @@ namespace PensionManagementBankingService.Controller
                 var updatedBankingDetails = await _bankingRepository.UpdateBankingDetailsById(bankId, bankDTO);
                 if(updatedBankingDetails == null)
                 {
-                    return NotFound();
+                    _logger.LogWarning($"No banking details found for ID: {bankId}");
+                    throw new BankingExceptions("No banking Details found for given Id");
 
                 }
+                _logger.LogInformation($"Updated BankingDetails for ID: {bankId}");
                 return Ok(updatedBankingDetails);
             }
             catch (BankingExceptions ex)
             {
-                switch (ex.Type)
-                {
-                    case BankingExceptions.ErrorType.DuplicateRecord:
-                        return Conflict(ex.Message);
-                    case BankingExceptions.ErrorType.EmptyResult:
-                        return NotFound(ex.Message);
-                    case BankingExceptions.ErrorType.NotFound:
-                        return NotFound(ex.Message);
-                    default:
-                        throw;
-                }
+                _logger.LogError(ex, $"Error occurred while updating banking details for given Id{bankId}");
+                return StatusCode(404,ex.Message);
             }
 
             catch (Exception ex)
             {
+                _logger.LogError(ex, $"An expected error occured while updating bank details for Id {bankId}");
                 return StatusCode(500, "An unexpected error occured, please try later");
             }
         }
         [HttpDelete("{bankId}")]
-        public IActionResult DeleteBankingDetailsById(Guid bankId)
+        public async Task<IActionResult> DeleteBankingDetailsById(Guid bankId)
         {
             try
             {
+                var existing = await _bankingRepository.GetBankingDetailsById(bankId);
+                if(existing == null)
+                {
+                    throw new BankingExceptions("Banking Details not found for given BankId");
+                }
                 _bankingRepository.DeleteBankingDetailsById(bankId);
+                
+                _logger.LogInformation($"Deleted BankingDetails for ID: {bankId}");
                 return Ok();
             }
             catch (BankingExceptions ex)
             {
-                switch (ex.Type)
-                {
-                    case BankingExceptions.ErrorType.DuplicateRecord:
-                        return Conflict(ex.Message);
-                    case BankingExceptions.ErrorType.EmptyResult:
-                        return NotFound(ex.Message);
-                    case BankingExceptions.ErrorType.NotFound:
-                        return NotFound(ex.Message);
-                    default:
-                        throw;
-                }
+                _logger.LogError(ex, $"Error occurred while deleting banking details for given Id{bankId}");
+                return NotFound(ex.Message);
             }
 
             catch (Exception ex)
             {
+                _logger.LogError(ex, $"An expected error occured while deleting bank details for Id {bankId}");
                 return StatusCode(500, "An unexpected error occured, please try later");
             }
         }
@@ -201,24 +187,21 @@ namespace PensionManagementBankingService.Controller
             try
             {
                 var result = await _bankingRepository.GetBankDetailsByPensionerId(pensionerId);
+                if(result == null)
+                {
+                    throw new BankingExceptions("BankId not found for given Pensioner Id");
+                }
+                _logger.LogInformation($"retrieved BankDetails for PensionerID: {pensionerId}");
                 return Ok(result);
             }
             catch(BankingExceptions ex)
             {
-                switch(ex.Type)
-                {
-                    case BankingExceptions.ErrorType.DuplicateRecord:
-                        return Conflict(ex.Message);
-                    case BankingExceptions.ErrorType.EmptyResult:
-                        return NotFound(ex.Message);
-                    case BankingExceptions.ErrorType.NotFound:
-                        return NotFound(ex.Message);
-                    default:
-                        throw;
-                }
+                _logger.LogError(ex, $"Error occurred while retrieving banking details for pensionerId{pensionerId}");
+                return StatusCode(404,ex.Message);
             }
             catch (Exception ex)
             {
+                _logger.LogError(ex, $"An expected error occured while retrieving bank details for PensionerId {pensionerId}");
                 return StatusCode(500, "An unexpected error occured, please try later");
             }
         }
